BUGFIX: In java .properties handler, utf-8 characters not displayed correctly in Lotte.

+++ transifex/resources/tests/lib/javaproperties/__init__.py
@@ -1,10 +1,10 @@
-import os
+import os, chardet
 import unittest
 from transifex.txcommon.tests.base import BaseTestCase
 from transifex.languages.models import Language
 from transifex.resources.models import *
 from transifex.resources.formats.javaproperties import  JavaPropertiesHandler, \
-        JavaParseError
+        JavaParseError, JavaCompileError
 
 from transifex.addons.suggestions.models import Suggestion
 
@@ -55,6 +55,22 @@ class TestJavaProperties(BaseTestCase):
         self.assertEqual(res[0], r"Key21\:WithColon")
         self.assertEqual(res[1], "Value21")
 
+    def test_convert_to_utf8(self):
+        handler = JavaPropertiesHandler(
+            os.path.join(os.path.dirname(__file__), 'complex_hi_IN-ascii.properties')
+        )
+        filename = handler.convert_to_utf8(handler.filename, handler.ENCODING)
+        content = open(filename, 'r').read()
+        if content.find('\\u') >= 0:
+            raise JavaParseError("File not properly converted to UTF-8")
+
+        handler.bind_file(
+            os.path.join(os.path.dirname(__file__), 'complex_hi_IN.properties')
+        )
+        self.assertRaises(JavaParseError, handler.convert_to_utf8,
+                handler.filename, handler.ENCODING)
+
+
     def test_properties_parser(self):
         """PROPERTIES file tests."""
         # Parsing PROPERTIES file
@@ -77,7 +93,7 @@ class TestJavaProperties(BaseTestCase):
         self.assertEqual(entities, 25)
         self.assertEqual(translations, 25)
 
-    def test_properties_save2db(self):
+    def test_properties_save2db(self, delete=True):
         """Test creating source strings from a PROPERTIES file works"""
         handler = JavaPropertiesHandler(
             os.path.join(os.path.dirname(__file__), 'complex.properties')
@@ -127,4 +143,21 @@ class TestJavaProperties(BaseTestCase):
         handler.bind_file(os.path.join(os.path.dirname(__file__),'complex_hi_IN.properties'))
         self.assertRaises(JavaParseError, handler.parse_file)
 
+        if delete:
+            r.delete()
+        else:
+            return r
+
+    def test_convert_to_ascii(self):
+        r = self.test_properties_save2db(delete=False)
+        handler = JavaPropertiesHandler()
+        handler.bind_resource(r)
+        handler.set_language(Language.objects.get(code='hi_IN'))
+        handler.compile()
+        compiled_template = handler.convert_to_ascii(handler.compiled_template)
+        if compiled_template.find('\\u') < 0:
+            raise JavaCompileError("Compiled template is not in %s format." %
+                    handler.ENCODING)
+        self.assertEqual(chardet.detect(compiled_template)['encoding'],
+                'ascii')
         r.delete()
